Move EWKB adapter defintions into ewkb macros
